+++ halfapi/__init__.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-__version__ = '0.6.22-rc2'
+__version__ = '0.6.22-rc3'
 
 def version():
     return f'HalfAPI version:{__version__}'